+++ android/library/src/main/java/com/liferay/mobile/screens/viewsets/defaultviews/ddl/form/fields/CustomRatingNumberView.java
@@ -97,10 +97,6 @@ public class CustomRatingNumberView extends LinearLayout
 	}
 
 	@Override
-	public void setPositionInParent(int position) {
-	}
-
-	@Override
 	public void setUpdateMode(boolean enabled) {
 		ratingBar.setEnabled(enabled);
 	}