+++ android/library/src/main/java/com/liferay/mobile/screens/viewsets/defaultviews/ddl/form/fields/DDLFieldNumberView.java
@@ -56,11 +56,6 @@ public class DDLFieldNumberView extends BaseDDLFieldTextView<NumberField> {
 	}
 
 	@Override
-	public void setPositionInParent(int position) {
-
-	}
-
-	@Override
 	protected void onTextChanged(String text) {
 		getField().setCurrentStringValue(text);
 	}