+++ android/library/src/main/java/com/liferay/mobile/screens/viewsets/defaultviews/ddl/form/fields/DDLFieldRadioView.java
@@ -126,11 +126,6 @@ public class DDLFieldRadioView extends RadioGroup
 	}
 
 	@Override
-	public void setPositionInParent(int position) {
-
-	}
-
-	@Override
 	public void setUpdateMode(boolean enabled) {
 		setEnabled(enabled);
 	}